+++ Assets/Scripts/DEBUG/MPSystem.cs
@@ -3,28 +3,30 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class MPBarSystem : MonoBehaviour
+public class MPSystem : MonoBehaviour
 {
     [SerializeField]
-    private Slider SliderCurrentMP;
+    private Text TextMaxMP;
+    [SerializeField]
+    private Text TextCurrentMP;
 
 
     public float MaxMP = 100.0f;
 
     public float MPChangeSpeed = 5.0f;
 
-    private ManaPoint MP;
+    private MP MP;
 
     private void Awake()
     {
-        MP = new ManaPoint(MaxMP);
-        SliderCurrentMP.maxValue = MaxMP;
+        MP = new MP(MaxMP);
     }
 
     private void Update()
     {
         MP.ChangeMP(Time.deltaTime * MPChangeSpeed);
-        SliderCurrentMP.value = MP.CurrentMP;
+        TextMaxMP.text = MP.MaxMP.ToString("N1");
+        TextCurrentMP.text = MP.CurrentMP.ToString("N1");
     }
 
     public void SpendMP(int damagePoint)
