+++ Assets/Scripts/Logic/HP.cs
@@ -1,72 +1,27 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
 
-public class HPSystem : MonoBehaviour
-{
-    [SerializeField]
-    private Text TextMaxHP;
-    [SerializeField]
-    private Text TextCurrentHP;
-    [SerializeField]
-    private Text TextFutureHP;
-
-    
-    public float MaxHP = 100.0f;
-
-    public float HPChangeSpeed = 10.0f;
-
-    private HitPoint HP;
-
-    private void Awake()
-    {
-        HP = new HitPoint(MaxHP);
-    }
-
-    private void Update()
-    {
-        HP.PersistentHP(Time.deltaTime * HPChangeSpeed);
-        TextMaxHP.text     = HP.MaxHP.ToString("N1");
-        TextFutureHP.text  = HP.FutureHP.ToString("N1");
-        TextCurrentHP.text = HP.CurrentHP.ToString("N1");
-    }
-
-    public void Damage(int damagePoint)
-    {
-        HP.ChangeHP(-damagePoint);
-    }
-
-    public void Heal(int healPoint)
-    {
-        HP.ChangeHP(healPoint);
-    }
-
-    public void Won()
-    {
-        HP.HPWhenWinning();
-    }
-}
-
-public class HitPoint
+public class HP
 {
     private float _currentHP;
-    public float CurrentHP {
+    public float CurrentHP
+    {
         get { return _currentHP; }
     }
 
     private float _futureHP;
-    public float FutureHP {
+    public float FutureHP
+    {
         get { return _futureHP; }
     }
 
     private float _maxHP;
-    public float MaxHP {
+    public float MaxHP
+    {
         get { return _maxHP; }
     }
 
 
-    public HitPoint(float maxHP)
+    public HP(float maxHP)
     {
         _maxHP = maxHP;
         _futureHP = maxHP;
