+++ Assets/Scripts/Logic/HP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db91656b69997344fa17833576ee706d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
