+++ Assets/Scripts/UnityObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72bd9545bab013545be77ada8522f8f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
